+++ src/SensorManager/Microphone.cpp
@@ -58,7 +58,7 @@ void Microphone::start(int sample_rate_idx) {
 
 	if (!_active) return;
 
-	LOG_INF("Starting Microphone at %f Hz", sample_rates.sample_rates[sample_rate_idx]);
+	LOG_INF("Starting Microphone at %f Hz", (double)sample_rates.sample_rates[sample_rate_idx]);
 	record_to_sd(_sd_logging);
 
 	audio_datapath_aquire(&fifo_rx);
